+++ lib/test_prof.rb
@@ -12,5 +12,6 @@ require "test_prof/factory_prof"
 require "test_prof/memory_prof"
 require "test_prof/rspec_stamp"
 require "test_prof/tag_prof"
+require "test_prof/tps_prof"
 require "test_prof/rspec_dissect" if TestProf.rspec?
 require "test_prof/factory_all_stub"